+++ roles/views.py
@@ -1,22 +1,27 @@
 # roles/views.py
 
 from django.shortcuts import render, redirect, get_object_or_404
-from django.contrib.auth.decorators import user_passes_test
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
 from .models import Role
 
-def is_admin(user):
-    """
-    Función que comprueba si un usuario es autenticado y de tipo ADMIN.
-    """
-    return user.is_authenticated and hasattr(user, 'tipo_usuario') and user.tipo_usuario == 'ADMIN'
 
-@user_passes_test(is_admin)
+# --- Función auxiliar para verificar si es admin ---
+def es_admin(user):
+    return user.is_authenticated and user.is_staff
+
+
+@login_required
 def role_panel(request):
     """
     Renderiza la página para administrar Roles y maneja la creación de nuevos roles.
+    Solo accesible para usuarios autenticados y staff.
     """
-    # Si el método es POST, significa que se está enviando el formulario para crear un rol.
+    if not request.user.is_staff:
+        messages.error(request, "No tienes permiso para acceder a esta página.")
+        return redirect("home")
+
     if request.method == 'POST':
         nombre_rol = request.POST.get('nombre')
         descripcion_rol = request.POST.get('descripcion')
@@ -27,20 +32,25 @@ def role_panel(request):
         else:
             messages.error(request, "El nombre del rol no puede estar vacío.")
         
-        return redirect('roles:role-panel') # Redirigimos para evitar reenvío del formulario
+        return redirect('roles:role-panel')  # Redirigimos para evitar reenvío del formulario
 
-    # Si el método es GET, simplemente mostramos la página con la lista de roles.
     roles = Role.objects.all()
     context = {
         'roles': roles
     }
     return render(request, 'roles/role_admin.html', context)
 
-@user_passes_test(is_admin)
+
+@login_required
 def role_delete(request, pk):
     """
     Vista para eliminar un rol específico.
+    Solo accesible para usuarios autenticados y staff.
     """
+    if not request.user.is_staff:
+        messages.error(request, "No tienes permiso para acceder a esta página.")
+        return redirect("home")
+
     rol_a_eliminar = get_object_or_404(Role, pk=pk)
     rol_nombre = rol_a_eliminar.nombre
     rol_a_eliminar.delete()
